+++ src/qa-bot.jsx
@@ -1,10 +1,10 @@
-import { Component } from "preact";
-import { ErrorBoundary, lazy } from "preact-iso";
+import { Component, lazy } from "react";
+import { ErrorBoundary } from "react-error-boundary";
 
 const AsyncLoadedQABot = lazy(() =>
-  import("@snf/access-qa-bot").then(module => ({
-    default: module.QABot
-  }))
+  import("@snf/access-qa-bot").then((module) => ({
+    default: module.QABot,
+  })),
 );
 
 export class QABot extends Component {
@@ -16,7 +16,11 @@ export class QABot extends Component {
       onOpenChange,
       apiKey,
       embedded,
-      loginUrl
+      loginUrl,
+      userEmail,
+      userName,
+      accessId,
+      onAnalyticsEvent
     } = this.props;
 
     // Support environment variable if apiKey is not provided via props
@@ -28,9 +32,10 @@ export class QABot extends Component {
     }
 
     // Detect login state if not provided via props
-    const loggedIn = isLoggedIn !== undefined
-      ? isLoggedIn
-      : document.cookie.split("; ").includes("SESSaccesscisso=1");
+    const loggedIn =
+      isLoggedIn !== undefined
+        ? isLoggedIn
+        : document.cookie.split("; ").includes("SESSaccesscisso=1");
 
     return (
       <ErrorBoundary>
@@ -42,6 +47,10 @@ export class QABot extends Component {
           embedded={embedded === true}
           apiKey={botApiKey}
           loginUrl={loginUrl}
+          userEmail={userEmail}
+          userName={userName}
+          accessId={accessId}
+          onAnalyticsEvent={onAnalyticsEvent}
         />
       </ErrorBoundary>
     );