+++ src/utils.jsx
@@ -1,5 +1,6 @@
-import { render } from "preact";
-import { useEffect, useMemo, useState } from "preact/hooks";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { useEffect, useMemo, useState } from "react";
 
 export const getMode = (breakpoint = 900) =>
   document.body.clientWidth >= breakpoint ? "desktop" : "mobile";
@@ -13,15 +14,14 @@ export const useMode = (breakpoint = 900) => {
 };
 
 export const renderShadow = (content, target, styles = []) => {
-  const shadow = target.attachShadow({ mode: "open" });
-  render(
-    <>
+  const shadow = createRoot(target.attachShadow({ mode: "open" }));
+  shadow.render(
+    <StrictMode>
       {content}
       {styles.map((style) => (
-        <style>{style}</style>
+        <style key={style}>{style}</style>
       ))}
-    </>,
-    shadow
+    </StrictMode>,
   );
 };
 
@@ -35,12 +35,27 @@ const fetchJsonDataCache = {};
 const fetchJsonPromiseCache = {};
 const fetchJson = (
   uri,
-  { cache = true, corsProxy = false } = { cache: true, corsProxy: false }
+  {
+    body = null,
+    cache = true,
+    corsProxy = false,
+    headers = undefined,
+    method = "GET",
+  } = {
+    body: null,
+    cache: true,
+    corsProxy: false,
+    headers: undefined,
+    method: "GET",
+  },
 ) => {
-  if (cache && uri in fetchJsonDataCache) return fetchJsonDataCache[uri];
-  if (uri in fetchJsonPromiseCache) return fetchJsonPromiseCache[uri];
+  const cacheKey = `${uri}|${body}`;
+  if (cache && cacheKey in fetchJsonDataCache)
+    return fetchJsonDataCache[cacheKey];
+  if (cacheKey in fetchJsonPromiseCache) return fetchJsonPromiseCache[cacheKey];
   const promise = fetch(
-    corsProxy ? `https://corsproxy.io/?${encodeURIComponent(uri)}` : uri
+    corsProxy ? `https://corsproxy.io/?${encodeURIComponent(uri)}` : uri,
+    { body, headers, method },
   ).then((response) => {
     delete fetchJsonPromiseCache[uri];
     if (response.status < 200 || response.status > 299) {
@@ -53,17 +68,27 @@ const fetchJson = (
       }
     }
   });
-  fetchJsonPromiseCache[uri] = promise;
+  fetchJsonPromiseCache[cacheKey] = promise;
   return promise;
 };
 
 export const useJSON = (
   uris,
-  { cache = true, corsProxy = false, defaultValue = null } = {
+  {
+    body = null,
+    cache = true,
+    corsProxy = false,
+    defaultValue = null,
+    headers = undefined,
+    method = "GET",
+  } = {
+    body: null,
     cache: true,
     corsProxy: false,
     defaultValue: null,
-  }
+    headers: undefined,
+    method: "GET",
+  },
 ) => {
   const [value, setValue] = useState(defaultValue);
   let single = false;
@@ -75,19 +100,21 @@ export const useJSON = (
     if (uris.length) {
       (async () => {
         const json = await Promise.all(
-          uris.map((uri) => fetchJson(uri, { cache, corsProxy }))
+          uris.map((uri) =>
+            fetchJson(uri, { body, cache, corsProxy, headers, method }),
+          ),
         );
         const result = single ? json[0] : json;
         setValue(result);
       })();
     }
-  }, [...uris]);
+  }, [uris.join(" ")]);
   return value;
 };
 
 export const useResourceGroups = () => {
   const response = useJSON(
-    "https://operations-api.access-ci.org/wh2/cider/v1/access-active-groups/type/resource-catalog.access-ci.org/"
+    "https://operations-api.access-ci.org/wh2/cider/v1/access-active-groups/type/resource-catalog.access-ci.org/",
   );
   return useTransform([response], ({ results }) => {
     const {
@@ -173,7 +200,7 @@ export const useResourceGroups = () => {
             .map((orgId) => orgId * -1)
             .filter((tagId) => tagIds.includes(tagId)),
           ...group.rollup_feature_ids.filter((featureId) =>
-            tagIds.includes(featureId)
+            tagIds.includes(featureId),
           ),
         ],
         resourceCategoryId: group.rollup_feature_ids.includes(137) ? 2 : 1,
@@ -182,10 +209,29 @@ export const useResourceGroups = () => {
           .map((orgId) => organizationsMap[orgId])
           .filter((org) => org),
         resources: resources.filter((res) =>
-          group.rollup_info_resourceids.includes(res.info_resourceid)
+          group.rollup_info_resourceids.includes(res.info_resourceid),
         ),
-      }))
-      .sort((a, b) => a.name.localeCompare(b.name));
+      }));
+
+    // Add aliases for CloudBank resource group.
+    const cloudBank = resourceGroups.find(
+      (group) => group.infoGroupId === "cloudbank.access-ci.org",
+    );
+    if (cloudBank)
+      for (const alias of [
+        "Amazon Web Services",
+        "Google Cloud",
+        "IBM Cloud",
+        "Microsoft Azure",
+      ])
+        resourceGroups.push({
+          ...cloudBank,
+          infoGroupId: `${alias.toLowerCase().replaceAll(" ", "-")}${cloudBank.infoGroupId}`,
+          parentInfoGroupId: cloudBank.infoGroupId,
+          name: alias,
+        });
+
+    resourceGroups.sort((a, b) => a.name.localeCompare(b.name));
 
     return linkResourceGroups({
       resourceCategories,
@@ -207,7 +253,7 @@ export const useResourceGroup = (infoGroupId, defaultValue = null) => {
 
 export const useResourceGroupResources = (infoGroupId, defaultValue = []) => {
   const res = useJSON(
-    `https://operations-api.access-ci.org/wh2/cider/v1/access-active/info_groupid/${infoGroupId}/?format=json`
+    `https://operations-api.access-ci.org/wh2/cider/v1/access-active/info_groupid/${infoGroupId}/?format=json`,
   );
   return res && !res.error ? res.results : defaultValue;
 };
@@ -215,7 +261,7 @@ export const useResourceGroupResources = (infoGroupId, defaultValue = []) => {
 export const useTransform = (
   responseArray,
   transformFunction,
-  defaultValue = null
+  defaultValue = null,
 ) => {
   return useMemo(() => {
     if (!responseArray.length) return defaultValue;
@@ -263,7 +309,7 @@ const linkResourceGroups = ({
 
 export const filterResourceGroups = (
   { resourceGroups, tags, tagCategories },
-  activeTagIds
+  activeTagIds,
 ) => {
   const resourceCategoryIds = new Set();
   const infoGroupIds = new Set();
@@ -274,7 +320,7 @@ export const filterResourceGroups = (
     if (tagIds.has(tag.tagId)) tagCategoryIds.add(tag.tagCategoryId);
 
   const activeTagCategories = tagCategories.filter(({ tagCategoryId }) =>
-    tagCategoryIds.has(tagCategoryId)
+    tagCategoryIds.has(tagCategoryId),
   );
 
   // TODO: Improve logic for disabled tags. Tags should be disabled if the active
@@ -284,7 +330,7 @@ export const filterResourceGroups = (
   const disabledTagIds = new Set(
     tags
       .filter((tag) => !tagCategoryIds.has(tag.tagCategoryId))
-      .map((tag) => tag.tagId)
+      .map((tag) => tag.tagId),
   );
 
   for (let resourceGroup of resourceGroups) {
@@ -319,7 +365,7 @@ export const filterResourceGroups = (
   };
 };
 
-export const sortOn = (prop) => (a, b) => a[prop] < b[prop] ? -1 : 1;
+export const sortOn = (prop) => (a, b) => (a[prop] < b[prop] ? -1 : 1);
 
 export const extractHref = (html) => {
   let href = html.match(/href="([^"]+)"/);
@@ -345,7 +391,7 @@ export const htmlToJsx = (html) => {
         if (href) paragraph.push(<a href={href}>{link[2]}</a>);
       }
     }
-    paragraphs.push(<p>{paragraph}</p>);
+    paragraphs.push(<p key={pText}>{paragraph}</p>);
   }
 
   return paragraphs;